Changes to the plugin interface:  - SimplePluginBase deprecated, it wasn't simple  - Plugins that relied on SimplePluginBase rewritten  - Properties test set now passed to GrinderPlugin.initialize  - Test.getNumber returns an int instead of an Integer

+++ src/net/grinder/console/swingui/TestGraphPanel.java
@@ -50,7 +50,7 @@ public class TestGraphPanel extends JPanel
 	while (testIterator.hasNext()) {
 	    final Test test = (Test)testIterator.next();
 
-	    final int testNumber = test.getTestNumber().intValue();
+	    final int testNumber = test.getTestNumber();
 	    final String description = test.getDescription();
 
 	    String label = testLabel + testNumber;
